New init script for local env to register / setup everything
